+++ geotrek/core/sql/40_troncons.sql
@@ -101,7 +101,7 @@ BEGIN
     FOR eid IN SELECT e.id
                FROM e_r_evenement_troncon et, e_t_evenement e
                WHERE et.troncon = NEW.id AND et.evenement = e.id
-               GROUP BY e.id
+               GROUP BY e.id, e.decallage
                HAVING BOOL_OR(et.pk_debut != et.pk_fin) OR e.decallage = 0.0
     LOOP
         PERFORM update_geometry_of_evenement(eid);
@@ -111,7 +111,7 @@ BEGIN
     FOR eid, egeom IN SELECT e.id, e.geom
                FROM e_r_evenement_troncon et, e_t_evenement e
                WHERE et.troncon = NEW.id AND et.evenement = e.id
-               GROUP BY e.id
+               GROUP BY e.id, e.geom, e.decallage
                HAVING COUNT(et.id) = 1 AND BOOL_OR(et.pk_debut = et.pk_fin) AND e.decallage != 0.0
     LOOP
         SELECT * INTO linear_offset, side_offset FROM ST_InterpolateAlong(NEW.geom, egeom) AS (position float, distance float);