+++ docs/changelog.rst
@@ -7,6 +7,9 @@ CHANGELOG
 
 **Bug fixes**
 
+* Refactor signals pre / post migrate according Django 1.8
+* Update translations
+* Fix path splitting
 * Fix AutoLogin Middleware with mapentity 3.1.4